+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - disable filter pin on context change
 - fix node status extraction
+- do not send default terminal size in shell and attach views
 
 ## 0.4.8 - 2026-05-04